Added debuging script to keep pytest compatibility
